jetson-nano-qspi-sd: Fix flashing Nano SD-CARD
Jetson Nano had bootloaders and firmwares relocated
to the qspi flash starting with L4T 32.5.1. We need
to keep the same layout and offsets for the SD-CARD,
and make sure all bootloaders and firmware go to
the boot flash.

From this commit on, only L4T 32.5.1 or newer
can be flashed. This corresponds to BalenaOS
v2.83.11 and newer.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>