assets: Set AGX Orin device specific size and partition sizes
... to match the cloud production release.

The staging device specific offset and root partition
size are set for the pre-production, (staging) release.

In this commit we set the production device specific size, where
the nvidia partitions are located, as well as the balena
root file sysitem partition size to match the ones
in a cloud image.

Change-type: patch
Signed-off-by: Alexandru Costache <alexandru@balena.io>