custom-xavier-nx-emmc: Add changes for example board addition
The uefi_jetson.bin binary is obtained from a local Yocto build,
from the image deploy directory.

The BSP archive should be served by a local webserver, but the
upstream BSP link can also be used.

This commit is incomplete because it does not include an updated XML file
with esp and other partitions added by L4T 35.X and should only
serve as an example for testing purposes.

Change-type: patch